Create object for getting environment variable from upstream
