+++ src/com/synopsys/integration/pipeline/scm/GitBranch.groovy
@@ -1,11 +1,7 @@
 package com.synopsys.integration.pipeline.scm
 
+import com.synopsys.integration.pipeline.jenkins.JenkinsUpstreamBuild
 import com.synopsys.integration.pipeline.jenkins.PipelineConfiguration
-import hudson.model.BuildListener
-import hudson.model.Cause
-import jenkins.model.Jenkins
-import org.jenkinsci.plugins.workflow.job.WorkflowJob
-import org.jenkinsci.plugins.workflow.job.WorkflowRun
 
 class GitBranch {
     private PipelineConfiguration pipelineConfiguration
@@ -19,49 +15,17 @@ class GitBranch {
     }
 
     String determineAndGetBranch() {
-        WorkflowRun currentBuild = pipelineConfiguration.getScriptWrapper().currentBuild().getRunWrapper().getRawBuild() as WorkflowRun
+        JenkinsUpstreamBuild jenkinsUpstreamBuild = new JenkinsUpstreamBuild(pipelineConfiguration)
+        String branchFromCause = jenkinsUpstreamBuild.getUpstreamEnvironmentVariable('BRANCH')
 
-        if (hasUpstreamCause(currentBuild)) {
-            String branchFromCause = determineBranchFromUpstreamCause(currentBuild)
-
-            if (branchFromCause?.trim()) {
-                setBranch(branchFromCause)
-                setBranchSource('upstream build')
-            }
+        if (branchFromCause?.trim()) {
+            setBranch(branchFromCause)
+            setBranchSource('upstream build')
         }
 
         pipelineConfiguration.getLogger().info("branch is set from ${branchSource} as ${branch}")
 
         return branch
-    }
-
-    private static boolean hasUpstreamCause(WorkflowRun workflowRun) {
-        return null != workflowRun.getCause(Cause.UpstreamCause)
-    }
-
-    private static String determineBranchFromUpstreamCause(WorkflowRun workflowRun) {
-        Cause.UpstreamCause upstreamCause = workflowRun.getCause(Cause.UpstreamCause)
-
-        Cause.UpstreamCause possiblyNull = getParentCause(upstreamCause)
-        while (possiblyNull != null) {
-            upstreamCause = possiblyNull
-            possiblyNull = getParentCause(possiblyNull)
-        }
-
-        WorkflowRun highestUpstreamBuild = getWorkflowRun(upstreamCause)
-        BuildListener buildListener = highestUpstreamBuild.getListener()
-
-        return upstreamCause.getUpstreamRun().getEnvironment(buildListener)['BRANCH']
-    }
-
-    private static Cause.UpstreamCause getParentCause(Cause.UpstreamCause cause) {
-        return getWorkflowRun(cause).getCause(Cause.UpstreamCause)
-    }
-
-    private static WorkflowRun getWorkflowRun(Cause.UpstreamCause cause) {
-        String jobName = cause.getUpstreamProject()
-        int buildNumber = cause.getUpstreamBuild()
-        return Jenkins.get().getItemByFullName(jobName, WorkflowJob).getBuildByNumber(buildNumber)
     }
 
     String getBranch() {
